delete mixin
